LayoutTests/svg/custom/display-none-a-does-not-stop-focus-navigation.html

<!DOCTYPE html> <!-- webkit-test-runner [ TabsToLinks=true ] -->
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<input id="input1"></input>
<svg display="none" width="100" height="100">
  <a xlink:href="#foo">
    <text y="20">Link</text>
  </a>
</svg>
<input id="input2"></input>
<script>
test(function() {
    input1.focus();
    assert_equals(document.activeElement, input1);
    eventSender.keyDown('\t');
    assert_equals(document.activeElement, input2);
}, 'SVG <a> which is display none does not block focus traversal');
</script>
